Fix round(double precision, int) in seed + KPI engine (Postgres has no such overload)
- 0004_seed.sql: cast sscore expression to numeric before round(...,2)
- 0005_kpi_engine.sql: cast execution_time_ms expression to numeric before round(...,2)

Surfaced when running against cloud Postgres (error 42883).

+++ supabase/migrations/0005_kpi_engine.sql
@@ -189,7 +189,7 @@ begin
     insert into kpi_runlog
       (kpi_id, tenant_id, input_records_count, calculation_version, execution_time_ms, status)
     values (d.id, p_tenant, nrec, 'v1',
-            round(extract(epoch from clock_timestamp() - t0) * 1000, 2), 'success');
+            round((extract(epoch from clock_timestamp() - t0) * 1000)::numeric, 2), 'success');
   end loop;
 end $$;
